IdleState

+++ Assets/Scripts/New/EnemyConfig.cs
@@ -5,7 +5,7 @@ using UnityEngine;
     [CreateAssetMenu(menuName = "EnemyConfig", fileName = "ScriptableObjects/EnemyConfig")]
     public class EnemyConfig : ScriptableObject
     {
-        public int Health;
+        public float Health;
         public float Speed;
         public int Damage;
         public GameObject Prefab; // Префаб врага
